+++ src/main/main.c
@@ -6,7 +6,7 @@
 /*   By: mfujishi <mfujishi@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/11 13:58:41 by mfujishi          #+#    #+#             */
-/*   Updated: 2022/06/14 15:59:18 by mfujishi         ###   ########.fr       */
+/*   Updated: 2022/06/14 22:34:18 by mfujishi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
